Using Unix EOL.

+++ daikin-zigbee-bridge/daikin-zigbee-bridge.ino
@@ -226,10 +226,10 @@ void debugMinMaxSetpoints() {
   int16_t actual_max_heat_setpoint      = 0;
   int16_t actual_abs_min_heat_setpoint  = 0;
   int16_t actual_abs_max_heat_setpoint  = 0;
-  bool success1 = zbThermostat.getGenericAttribute((uint16_t)ESP_ZB_ZCL_CLUSTER_ID_THERMOSTAT, (uint16_t)ESP_ZB_ZCL_ATTR_THERMOSTAT_MIN_HEAT_SETPOINT_LIMIT_ID,      actual_min_heat_setpoint    );
-  bool success2 = zbThermostat.getGenericAttribute((uint16_t)ESP_ZB_ZCL_CLUSTER_ID_THERMOSTAT, (uint16_t)ESP_ZB_ZCL_ATTR_THERMOSTAT_MAX_HEAT_SETPOINT_LIMIT_ID ,     actual_max_heat_setpoint    );
-  bool success3 = zbThermostat.getGenericAttribute((uint16_t)ESP_ZB_ZCL_CLUSTER_ID_THERMOSTAT, (uint16_t)ESP_ZB_ZCL_ATTR_THERMOSTAT_ABS_MIN_HEAT_SETPOINT_LIMIT_ID,  actual_abs_min_heat_setpoint);
-  bool success4 = zbThermostat.getGenericAttribute((uint16_t)ESP_ZB_ZCL_CLUSTER_ID_THERMOSTAT, (uint16_t)ESP_ZB_ZCL_ATTR_THERMOSTAT_ABS_MAX_HEAT_SETPOINT_LIMIT_ID,  actual_abs_max_heat_setpoint);
+  bool success1 = zbThermostat.getGenericAttribute(ESP_ZB_ZCL_CLUSTER_ID_THERMOSTAT, ESP_ZB_ZCL_ATTR_THERMOSTAT_MIN_HEAT_SETPOINT_LIMIT_ID,      actual_min_heat_setpoint    );
+  bool success2 = zbThermostat.getGenericAttribute(ESP_ZB_ZCL_CLUSTER_ID_THERMOSTAT, ESP_ZB_ZCL_ATTR_THERMOSTAT_MAX_HEAT_SETPOINT_LIMIT_ID ,     actual_max_heat_setpoint    );
+  bool success3 = zbThermostat.getGenericAttribute(ESP_ZB_ZCL_CLUSTER_ID_THERMOSTAT, ESP_ZB_ZCL_ATTR_THERMOSTAT_ABS_MIN_HEAT_SETPOINT_LIMIT_ID,  actual_abs_min_heat_setpoint);
+  bool success4 = zbThermostat.getGenericAttribute(ESP_ZB_ZCL_CLUSTER_ID_THERMOSTAT, ESP_ZB_ZCL_ATTR_THERMOSTAT_ABS_MAX_HEAT_SETPOINT_LIMIT_ID,  actual_abs_max_heat_setpoint);
   logEntry("success=%d  actual_min_heat_setpoint=%d",     success1, actual_min_heat_setpoint);
   logEntry("success=%d  actual_max_heat_setpoint=%d",     success2, actual_max_heat_setpoint);
   logEntry("success=%d  actual_abs_min_heat_setpoint=%d", success3, actual_abs_min_heat_setpoint);
